authorize endpoints and save user details to when creating job

+++ backend/job/views.py
@@ -2,13 +2,15 @@ from itertools import count
 from django.shortcuts import get_object_or_404
 from django.shortcuts import render
 from rest_framework.response import Response
-from rest_framework.decorators import api_view
+from rest_framework.decorators import api_view, permission_classes
 from rest_framework import status
 from rest_framework.pagination import PageNumberPagination
 from .serializers import JobSerializer
 from django.db.models import Avg, Min, Max, Count
 from .models import Job
 from .filters import JobsFilter
+
+from rest_framework.permissions import IsAuthenticated
 
 # Create your views here.
 
@@ -43,7 +45,9 @@ def getJob(request, pk):
 
 
 @api_view(['POST'])
+@permission_classes([IsAuthenticated])
 def newJob(request):
+    request.data['user'] = request.user
     data = request.data
     job = Job.objects.create(**data)
 
@@ -52,9 +56,13 @@ def newJob(request):
 
 
 @api_view(['PUT'])
+@permission_classes([IsAuthenticated])
 def updateJob(request, pk):
     # job = Job.objects.get(id=pk)
     job = get_object_or_404(Job, id=pk)
+
+    if job.user != request.user:
+        return Response({'message': 'You cannot update this job'}, status=status.HTTP_403_FORBIDDEN)
 
     job.title = request.data['title']
     job.description = request.data['description']
@@ -76,8 +84,13 @@ def updateJob(request, pk):
 
 
 @api_view(['DELETE'])
+@permission_classes([IsAuthenticated])
 def deleteJob(request, pk):
     job = get_object_or_404(Job, id=pk)
+
+    if job.user != request.user:
+        return Response({'message': 'You cannot delete this job'}, status=status.HTTP_403_FORBIDDEN)
+
     job.delete()
 
     return Response({'message': 'Job is deleted'}, status=status.HTTP_200_OK)
